refactor: extract i18n funciton
